requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.1
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.4.0
tiktoken>=0.5.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.6.0
types-requests>=2.31.0

# Enhanced OCR dependencies
paddlepaddle-cpu>=2.5.2
paddleocr>=2.7.0
FlagEmbedding>=1.2.0
opencv-python>=4.8.0  # For advanced OCR preprocessing

# Desktop launcher and system monitoring
psutil>=5.9.0
pystray>=0.19.0
pillow>=9.0.0

# Optional dependencies for enhanced features
GPUtil>=1.4.0  # For GPU monitoring
pynvml>=11.5.0  # Alternative GPU monitoring

# Chart generation for analytics
matplotlib>=3.6.0
plotly>=5.15.0

# Windows-specific dependencies (install only on Windows)
pywin32>=305  # For Windows registry and COM operations
winshell>=0.6  # For Windows shell operations

# Linux-specific dependencies
# python-xlib>=0.33  # For X11 operations on Linux

# RAG v2 Advanced Dependencies
langchain>=0.2.4
langgraph>=0.0.30
auto-gptq>=0.7.0
torch==2.2.1
accelerate>=0.21.0
transformers>=4.36.0
bitsandbytes>=0.41.0
ragas==0.0.20
pypdf==4.2.0
optimum>=1.13.0
sentencepiece>=0.1.99
safetensors>=0.3.1

# Additional v2 dependencies
uvloop>=0.19.0  # High-performance event loop
websockets>=12.0  # WebSocket streaming support 